backend/services/payment_service/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.10
pika==1.3.2
redis==5.0.4
dj-database-url==2.1.0
gunicorn==22.0.0
requests==2.31.0
python-dotenv==1.0.1